Different Strafing Algo
I have added increased functionality where you can now toggle between auto strafing with "t".

This algorithm measure the time between a press and a release of a button instead of measuring the time between subsequent presses.

I have also examined the effects of different press times in game and correlated the press times with the characters momentum. Right now the code works pretty good but further refinements of the code are needed.